Isolated In Game UI Code, Beginning Viewport implementation
Pulled in-game ui code into game_ui.py to help readability
Began implementing basic viewport styled camera. Able to build a camera with a set size and pass it the main and surrounding chunks each update call.
Next steps are to finalize tile coordinates, give the camera a temp position to manipulate, and selectively grab tiles in view to create the camera's map string to render.