Load Menu Buttons
Load menu queries the DB for saves and assigns each a button. The buttons are capable of swapping to the game_ui. The GameManagementSystem has also been moved to the game_ui. The buttons need functionality to pass the world they posses to the Management System when changing to the game ui